+++ src/main/java/com/portiony/portiony/PortionyApplication.java
@@ -1,4 +1,4 @@
-package com.portiony.portiony.service;
+package com.portiony.portiony;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
